+++ Emu/source/Camera/ScrollingCamera.cpp
@@ -7,7 +7,7 @@ namespace Engine
 {
 	ScrollingCamera::ScrollingCamera() : m_scrollSpeed(0, 0), Camera() {}
 
-	void ScrollingCamera::Update(const float interpolation)
+	void ScrollingCamera::Update(const double interpolation)
 	{
 		m_offset += m_scrollSpeed;
 		if (m_clampingOn) Clamp();
